requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
selenium-stealth==1.0.6
python-dotenv==1.0.1
